pyproject.toml
[tool.poetry]
name = "notion-todoist-sync"
version = "0.1.0"
description = "Sync Notion tasks to Todoist"
authors = ["Meng Pan"]

[tool.poetry.dependencies]
python = "^3.9"
notion-client = "^2.0.0"
todoist-api-python = "^3.2.1"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
aiohttp = "^3.9.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sync = "notion_todoist_sync.sync_runner:main"
webhook-server = "notion_todoist_sync.webhook_server:main"
